+++ x/btcstaking/types/query.go
@@ -81,5 +81,6 @@ func NewFinalityProviderResponse(f *FinalityProvider, bbnBlockHeight uint64) *Fi
 		SlashedBtcHeight:     f.SlashedBtcHeight,
 		Jailed:               f.Jailed,
 		Height:               bbnBlockHeight,
+		HighestVotedHeight:   f.HighestVotedHeight,
 	}
 }